Change fix to instead move the TestStream and WatermarkController implementation to the runner
